% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{backtesting}
\alias{backtesting}
\title{backtesting}
\usage{
backtesting(cash, amount_per, date, signal, stock, buy_signal, sell_signal)
}
\arguments{
\item{cash}{交易开始时持有的现金}

\item{amount_per}{每次执行策略所买入的金额}

\item{date}{交易向量}

\item{signal}{交易信号}

\item{stock}{交易标的}

\item{buy_signal}{买入信号}

\item{sell_signal}{卖出信号}
}
\value{
返回一个数据框，date列表示回测过程的日期，strategy表示执行策略收获的总资产，
baseline表示开始全仓买入收获的总资产
}
\description{
不考虑交易的各种手续费的简单策略回测函数。
当signal信号大于buy_signal时，买入amout_per金额的股票，
当signal信号小于sell_signal时，卖出上一次交易买入的股票数。
最终该函数将返回一个数据框，数据框包括整个回测过程的日期、信号交易以及全仓买入股票两种策略所持有的现金和股票的总价值。
}
