% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE.R
\name{CASE}
\alias{CASE}
\title{CASE Model}
\usage{
CASE(
  Z = NULL,
  R,
  N,
  hatB = NULL,
  hatS = NULL,
  V = NULL,
  h = NULL,
  cs = TRUE,
  pi.init = NULL,
  U.orig = NULL,
  V.fix = TRUE,
  pi.fix = FALSE,
  n.iter = 45,
  MC.max = 125,
  tol = 0.01,
  MC.seq = "exp",
  MC.sim = 41,
  MC.sample = 58,
  significant_thres = 0.1,
  noise_threshold = 0.1
)
}
\arguments{
\item{Z}{M * C matrix of z scores.}

\item{R}{M * M matrix of LD.}

\item{hatB}{M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.}
}
\value{
A \code{"CASE"} object with the following elements:
\item{pi:}{L-vector, the prior probabilities of sharing patterns.}
\item{U:}{L-list of C * C matrix, the prior covariances of sharing patterns.}
\item{V:}{C * C matrix, the sample-adjusted phenotypical variance.}
}
\description{
Perform Multi-trait Fine-mapping
}
\details{
TBD
}
\examples{
TBD
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
