% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE_models.R
\name{CASE_test}
\alias{CASE_test}
\title{CASE Model Testing}
\usage{
CASE_test(
  Z = NULL,
  R,
  hatB = NULL,
  hatS = NULL,
  N,
  CASE_training,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{Z}{M * C matrix of z scores.}

\item{R}{M * M matrix of LD.}

\item{hatB}{M * C matrix of the estimated effects. Alternative summary data (together with hatS) to be provided instead of Z.}

\item{hatS}{M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.}

\item{N}{either 1 or C vector of the sample size, or C * C matrix of the sample size (diagonal) and overlaps  (off-diagonal). If provided with a vector, CASE assumes that each pair of traits overlaps with their minimal sample size.}

\item{CASE_training}{A \code{"CASE_training"} object.}

\item{verbose}{(optional) logical, whether to print logging information. Default = TRUE.}

\item{...}{additional arguments.}
}
\value{
A \code{"CASE"} object with the following elements:
\item{pi:}{L-vector, the prior probabilities of sharing patterns.}
\item{U:}{L-list of C * C matrix, the prior covariances of sharing patterns.}
\item{V:}{C * C matrix, the sample-adjusted phenotypical variance.}
\item{pip:}{M * C matrix, posterior probability of having eQTL effects per SNP per cell type.}
\item{post_mean:}{M * C matrix, average posterior estimates of eQTL effects per SNP per cell type.}
}
\description{
Obtain the posterior probabilities and estimates of the cis-eQTL effect sizes.
}
\details{
TBD
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
