% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE_models.R
\name{CASE_train}
\alias{CASE_train}
\title{CASE Model Training}
\usage{
CASE_train(Z = NULL, R, hatB = NULL, hatS = NULL, N, V = NULL, ...)
}
\arguments{
\item{Z}{M * C matrix of z scores.}

\item{R}{M * M matrix of LD.}

\item{hatB}{M * C matrix of the estimated effects. Alternative summary data (together with hatS) to be provided instead of Z.}

\item{hatS}{M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.}

\item{N}{either C vector of the sample size, or C * C matrix of the sample size (diagonal) and ovelaps  (off-diagonal). If provided with a vector, CASE assumes that each pair of traits overlaps with their minimal sample size.}

\item{V}{(optional) C * C covariance (correlation) matrix for the noise between traits. If not provided, the default is an identity matrix.}

\item{...}{additional arguments.}
}
\value{
A \code{"CASE_training"} object with the following elements:
\item{pi:}{L-vector, the prior probabilities of sharing patterns.}
\item{U:}{L-list of C * C matrix, the prior covariances of sharing patterns.}
\item{V:}{C * C matrix, the sample-adjusted phenotypical variance.}
}
\description{
Fit the priors for the cis-eQTL effect sizes.
}
\details{
TBD
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
