% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE1.0.R
\name{ED_test}
\alias{ED_test}
\title{CASE Model Testing}
\usage{
ED_test(hatB = NULL, Z = NULL, R, N, V, U, pi, MC.sim = 41, MC.sample = 58)
}
\arguments{
\item{Z}{(M * C),The z scores for eQTL summary statistics.}
}
\value{
The temperature in degrees Celsius
}
\description{
Obtain the posterior probabilities and estimates of the cis-eQTL effect sizes.
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
