% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
## `example_data`
A list contains 3 elements:
\describe{
  \item{Y}{500 by 3 matrix of phenotype.}
  \item{X}{500 by 1000 matrix of genotype. This is created by R package `CorBin`}
  \item{B}{1000 by 3 matrix of eQTL effects. Only the 10th and 950th SNPs have eQTL effects.}
}
}
\usage{
example_data
}
\description{
An example data of three traits for the CASE fine-mapping illustration.
}
\keyword{datasets}
