% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/garch.R
\name{garch}
\alias{garch}
\title{GARCH}
\usage{
garch(data, fore_L, tau)
}
\arguments{
\item{data}{数据框，每一列代表一天的价格数据}

\item{fore_L}{选取data数据框的尾部fore_L列作为预测对象}

\item{tau}{计算VaR时的选择的分位数}
}
\value{
返回一个数据框，每一列依次为:收益率预测、波动率预测、基于garch(1,1)得到的VaR预测
}
\description{
基于garch(1,1)模型和arma(1,1)模型得到的波动率、收益率以及VaR预测的函数
}
