% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har_rv.R
\name{har_rv}
\alias{har_rv}
\title{har_rv}
\usage{
har_rv(data, fore_L, tau)
}
\arguments{
\item{data}{数据框，每一列代表一天的价格数据}

\item{fore_L}{选取data数据框的尾部fore_L列作为预测对象}

\item{tau}{计算VaR时的选择的分位数}
}
\value{
返回一个数据框，每一列依次为:收益率预测、波动率预测、基于HAR-RV方法得到的VaR预测以及
基于历史模拟法得到的VaR的预测
}
\description{
har_rv
}
