% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iVXCalc.R
\name{iVXCalc}
\alias{iVXCalc}
\title{iVXCalc}
\usage{
iVXCalc(data, rate)
}
\arguments{
\item{data}{数据框，要求包含期权数据的交易日期、期权合约、类型、挂牌日期以及收盘价等数据}

\item{rate}{与期权数据长度相等的利率向量或者常数，默认为利率为常数0.04}
}
\value{
返回估计得到的iVX数据框，包括iVX的日期和值
}
\description{
用于估计中国波指iVX
}
