% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moutai_data.R
\docType{data}
\name{moutai_data}
\alias{moutai_data}
\title{该数据集为贵州茅台自2012年至2023年的五分钟高频股票价格数据
去除了原始价格数据中的午餐效应、隔夜效应以及缺失值
数据集类型为数据框，每一列代表一天内的价格变动情况}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 49 rows and 2718 columns.
}
\source{
\url{https://www.resset.com/}
}
\usage{
moutai_data
}
\description{
该数据集为贵州茅台自2012年至2023年的五分钟高频股票价格数据
去除了原始价格数据中的午餐效应、隔夜效应以及缺失值
数据集类型为数据框，每一列代表一天内的价格变动情况
}
\keyword{datasets}
