% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_data.R
\docType{data}
\name{option_data}
\alias{option_data}
\title{option data}
\format{
\describe{
\item{交易日期}
\item{类型}
\item{到期日期}
\item{执行价}
\item{收盘价}
}
}
\source{
\url{https://www.joinquant.com/}
}
\usage{
option_data
}
\description{
包含了2015年2月至2024年5月的288756条期权数据
}
\keyword{datasets}
