% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE.R
\name{CASE}
\alias{CASE}
\title{CASE}
\usage{
CASE(Z = NULL, R, hatB = NULL, hatS = NULL, N, V = NULL, cs = TRUE, ...)
}
\arguments{
\item{Z}{M * C matrix of z scores.}

\item{R}{M * M matrix of LD.}

\item{hatB}{M * C matrix of the estimated effects. Alternative summary data (together with hatS) to be provided instead of Z.}

\item{hatS}{M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.}

\item{N}{either C vector of the sample size, or C * C matrix of the sample size (diagonal) and ovelaps  (off-diagonal). If provided with a vector, CASE assumes that each pair of traits overlaps with their minimal sample size.}

\item{V}{(optional) C * C covariance (correlation) matrix for the noise between traits. If not provided, the default is an identity matrix.}

\item{cs}{logical, whether to get credible sets.}

\item{...}{additional arguments.}
}
\value{
A \code{"CASE"} object with the following elements:
\item{pi:}{L-vector, the prior probabilities of sharing patterns.}
\item{U:}{L-list of C * C matrix, the prior covariances of sharing patterns.}
\item{V:}{C * C matrix, the sample-adjusted phenotypical variance.}
}
\description{
Perform Multi-trait Fine-mapping
}
\details{
TBD
}
\examples{
## A single-trait example.
set.seed(3)
N = 500
M = 1000
X = matrix(sample(0:2, size = N * M, replace = TRUE), N, M)
B = rep(0, M)
B[1:4] = runif(4, 1, 2)
e = rnorm(N)
Y = X \%*\% B + e

X = scale(X)
Y = scale(Y)
R = cor(X)

Z = hatB = hatS = rep(0, M)
hatB <- as.vector(t(X) \%*\% Y / (N - 1))
hatS <- sqrt((sum(Y^2) / (N - 1)  - hatB^2 ) / (N - 1))
Z <- hatB / hatS

fit <- CASE(Z = Z, R = R, N = N)
# print(fit$sets)


## A multi-trait example.
set.seed(3)
N = 500
M = 1000
C = 3
X = matrix(sample(0:2, size = N * M, replace = TRUE), N, M)
B = matrix(0, M, C)
B[1, ] = runif(C, 1, 2)
B[2, 1] = runif(1, 2, 3)
e = matrix(rnorm(N * C), N, C)
Y = X \%*\% B + e

X = scale(X)
Y = scale(Y)
R = cor(X)

Z = hatB = hatS = matrix(0, M, C)
for (c in 1:C){
  hatB[, c] <- t(X) \%*\% Y[, c] / (N - 1)
  hatS[, c] <- sqrt((sum(Y[, c]^2) / (N - 1)  - hatB[, c]^2 ) / (N - 1))
  Z[, c] <- hatB[, c] / hatS[, c]
}

fit <- CASE(Z = Z, R = R, N = rep(N, C))
# print(fit$sets)
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
