% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE1.0.R
\name{ED_test}
\alias{ED_test}
\title{CASE Model Testing}
\usage{
ED_test(hatB = NULL, Z = NULL, R, N, V, U, pi, MC.sim = 41, MC.sample = 58)
}
\arguments{
\item{hatB}{M * C matrix of standard errors of the estimated effects. Alternative summary data (together with hatB) to be provided instead of Z.}

\item{Z}{M * C matrix of z scores.}

\item{R}{M * M matrix of LD.}
}
\value{
A \code{"CASE"} object with the following elements:
\item{pi:}{L-vector, the prior probabilities of sharing patterns.}
\item{U:}{L-list of C * C matrix, the prior covariances of sharing patterns.}
\item{V:}{C * C matrix, the sample-adjusted phenotypical variance.}
}
\description{
Obtain the posterior probabilities and estimates of the cis-eQTL effect sizes.
}
\details{
TBD
}
\examples{
TBD
}
\references{
TBD
}
\author{
Chen Lin, Hongyu Zhao
}
