% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CASE1.0.R
\name{get_credible_sets}
\alias{get_credible_sets}
\title{CASE Obtain Credible Sets}
\usage{
get_credible_sets(pvalues, R, cor.min = 0.5, pip = 0.95, ruled_out = 1 - 1e-04)
}
\arguments{
\item{pvalues}{(M * C),The pvalues of SNPs.}
}
\value{
Credible Sets
}
\description{
Obtain credible sets for any multi-trait fine-mapping results.
}
\examples{
TBD
}
